+++ R/data.R
@@ -1,0 +1,135 @@
+pride_data <- list(
+  # Base flags
+  rainbow = list(
+    colors = c("#E40303", "#FF8C00", "#FFED00", "#008026", "#24408E", "#732982")
+  ),
+  lesbian = list(
+    colors = c("#D62900", "#FF9B55", "#FFFFFF", "#D461A6", "#A50062")
+  ),
+  gay = list(
+    colors = c("#078D70", "#98E8C1", "#FFFFFF", "#7BADE2", "#3D1A78")
+  ),
+  bisexual = list(
+    colors = c("#D00070", "#8C4799", "#0032A0")
+  ),
+  transgender = list(
+    colors = c("#55CDFC", "#F7A8B8", "#FFFFFF")
+  ),
+  asexual = list(
+    colors = c("#000000", "#A4A4A4", "#FFFFFF", "#810081")
+  ),
+  intersex = list(
+    colors = c("#FFD800", "#7902AA")
+  ),
+  nonbinary = list(
+    colors = c("#FFF430", "#FFFFFF", "#9C59D1", "#000000")
+  ),
+  # Rainbow variants
+  philadelphia = list(
+    colors = c("#000000", "#784F17", "#D12229", "#F68A1E", "#FDE01A", "#007940", "#24408E", "#732982")
+  ),
+  progress = list(
+    colors = c("#D12229", "#F68A1E", "#FDE01A", "#007940", "#24408E", "#732982",
+               "#FFFFFF", "#FFAFC8", "#74D7EE", "#784F17", "#000000")
+  ),
+  # A-spectrum
+  aromantic = list(
+    colors = c("#3DA542", "#A7D379", "#FFFFFF", "#A4A4A4", "#000000")
+  ),
+  acesthetic = list(
+    colors = c("#000000", "#A4A4A4", "#FFFFFF", "#FD82AE")
+  ),
+  analterous = list(
+    colors = c("#000000", "#A4A4A4", "#FFFFFF", "#75A0FF")
+  ),
+  asensual = list(
+    colors = c("#000000", "#A4A4A4", "#FFFFFF", "#FFD698")
+  ),
+  ace_spectrum = list(
+    colors = c("#151747", "#7557AD", "#D992B4", "#FFEDED")
+  ),
+  aro_spectrum = list(
+    colors = c("#3FA053", "#96D07C", "#F3FCCF", "#176363", "#052027")
+  ),
+  aroace = list(
+    colors = c("#E28C00", "#ECCD00", "#FFFFFF", "#62AEDC", "#203856")
+  ),
+  alloace = list(
+    colors = c("#D50100", "#F85858", "#FFFFFF", "#81007F")
+  ),
+  aroallo = list(
+    colors = c("#3DA542", "#A8D379", "#FFFFFF", "#FBE34B", "#E8BB00")
+  ),
+  # Other sexualities
+  pansexual = list(
+    colors = c("#FF1B8D", "#FFDA00", "#1BB3FF")
+  ),
+  androsexual = list(
+    colors = c("#00CCFF", "#603524", "#B79ADD")
+  ),
+  gynesexual = list(
+    colors = c("#F3A9B6", "#903F2A", "#5A953B")
+  ),
+  abrosexual = list(
+    colors = c("#37CBBD", "#95E9B1", "#FFFFFF", "#FEA396", "#F35777")
+  ),
+  lesbian_7 = list(
+    colors = c("#A60061", "#B95594", "#D162A8", "#FFFFFF", "#E5ADD1", "#C64D53", "#8C1801")
+  ),
+  gay_7 = list(
+    colors = c("#078D70", "#26CEAA", "#98E8C1", "#FFFFFF", "#7BADE2", "#5049CC", "#3D1A78")
+  ),
+  # Fem-masc genders
+  demiboy = list(
+    colors = c("#7F7F7F", "#C4C4C4", "#9DD7EA", "#FFFFFF")
+  ),
+  demigirl = list(
+    colors = c("#7F7F7F", "#C4C4C4", "#FDADC8", "#FFFFFF")
+  ),
+  # Galactian alignment system
+  helian = list(
+    colors = c("#FF7367", "#BC3028", "#FFF7B8", "#000000", "#B5B5B5")
+  ),
+  lunarian = list(
+    colors = c("#C6A5D0", "#85618D", "#FFF7B8", "#000000", "#B5B5B5")
+  ),
+  solarian = list(
+    colors = c("#FFB187", "#FF7D51", "#FFF7B8", "#000000", "#B5B5B5")
+  ),
+  stellarian = list(
+    colors = c("#B6D6C1", "#79C49D", "#FFF7B8", "#000000", "#B5B5B5")
+  ),
+  # Other genders
+  genderfluid = list(
+    colors = c("#FF76A4", "#FFFFFF", "#C011D7", "#000000", "#2F3CBE")
+  ),
+  genderqueer = list(
+    colors = c("#B57EDC", "#FFFFFF", "#4A8123")
+  ),
+  agender = list(
+    colors = c("#000000", "#BCC4C7", "#FFFFFF", "#B7F684")
+  ),
+  bigender = list(
+    colors = c("#C479A2", "#EDA5CD", "#D6C7E8", "#FFFFFF", "#9AC7E8", "#6D82D1")
+  ),
+  # Subcultures
+  amatopunk = list(
+    colors = c("#531768", "#B74D71", "#DB966D", "#FFE7AA", "#4E8C69")
+  ),
+  bear = list(
+    colors = c("#623804", "#D56300", "#FEDD63", "#FEE6B8", "#FFFFFF", "#555555", "#000000")
+  ),
+  butch = list(
+    colors = c("#324070", "#6A7AA4", "#8A92A9", "#EDECEB", "#C0B4DB", "#764EC5", "#51048A")
+  ),
+  femme = list(
+    colors = c("#7D0037", "#DE42F1", "#FDFDFD", "#F29C41", "#9E470E")
+  ),
+  otter = list(
+    colors = c("#23A8FA", "#9AD7FC", "#FFFFFF", "#B4B4B4", "#2A2A2A")
+  ),
+  # Uncategorized
+  queerhet = list(
+    colors = c("#4E8492", "#6CCB99", "#727D6E", "#DAC686", "#B86962")
+  )
+)