+++ R/themes.R
@@ -5,25 +5,27 @@
 #' light. Conversely, this same gray is too dark for flags with yellow stripes.
 #' LGBTQ themes are carefully curated for each palette.
 #'
-#' @template name
-#' @param ... \code{ANY}\cr
-#'  Parameters passed to \code{\link[ggplot2]{theme}()}.
+#' @param name `character(1)` \cr
+#'  Name of the flag the colors are based on.
+#' @param ... `ANY` \cr
+#'  Parameters passed to [ggplot2::theme()].
 #'
 #' @eval roxygen_available_palettes()
 #'
 #' @return A ggplot2 `theme` object.
 #'
 #' @examples
-#' library(ggplot2)
-#' ggplot(data.frame(x = 1:10, y = 15:6,
-#'                   group = rep(c("a", "b"), each = 5)),
-#'        aes(x = x, y = y, color = group)) +
-#'   geom_point(size = 4) +
-#'   scale_color_manual(values = palette_lgbtq("intersex")) +
+#' data <- data.frame(
+#'   x = 1:10, y = 15:6,
+#'   group = rep(c("a", "b"), each = 5)
+#' )
+#'
+#' ggplot2::ggplot(data, ggplot2::aes(x = x, y = y, color = group)) +
+#'   ggplot2::geom_point(size = 4) +
+#'   scale_color_lgbtq("intersex") +
 #'   # Use the same name as for values, preferably
 #'   theme_lgbtq("intersex")
 #'
-#' @importFrom ggplot2 theme element_rect element_line rel
 #' @export
 theme_lgbtq <- function(name, ...) {
   if (!name %in% names(pride_data)) {
@@ -35,47 +37,47 @@ theme_lgbtq <- function(name, ...) {
 
 theme_lgbtq_default <- function(...) {
   ggplot2::theme(
-    panel.background = element_rect(fill = "grey82", color = NA),
-    panel.grid.minor = element_line(size = rel(0.5)),
-    legend.key = element_rect(fill = "grey82", color = NA),
+    panel.background = ggplot2::element_rect(fill = "grey82", color = NA),
+    panel.grid.minor = ggplot2::element_line(linewidth = ggplot2::rel(0.5)),
+    legend.key = ggplot2::element_rect(fill = "grey82", color = NA),
     ...
   )
 }
 
 theme_lgbtq_white <- function(...) {
   ggplot2::theme(
-    panel.background = element_rect(fill = "white", color = NA),
-    panel.border = element_rect(fill = NA, color = "black"),
-    panel.grid = element_line(color = "grey92"),
-    panel.grid.minor = element_line(size = rel(0.5)),
-    legend.key = element_rect(fill = "white", color = NA),
+    panel.background = ggplot2::element_rect(fill = "white", color = NA),
+    panel.border = ggplot2::element_rect(fill = NA, color = "black"),
+    panel.grid = ggplot2::element_line(color = "grey92"),
+    panel.grid.minor = ggplot2::element_line(linewidth = ggplot2::rel(0.5)),
+    legend.key = ggplot2::element_rect(fill = "white", color = NA),
     ...
   )
 }
 
 theme_agender <- function(...) {
   ggplot2::theme(
-    panel.background = element_rect(fill = "#DDE3D8", color = NA),
-    panel.grid.minor = element_line(size = rel(0.5)),
-    legend.key = element_rect(fill = "#DDE3D8", color = NA),
+    panel.background = ggplot2::element_rect(fill = "#DDE3D8", color = NA),
+    panel.grid.minor = ggplot2::element_line(linewidth = ggplot2::rel(0.5)),
+    legend.key = ggplot2::element_rect(fill = "#DDE3D8", color = NA),
     ...
   )
 }
 
 theme_demiboy <- function(...) {
   ggplot2::theme(
-    panel.background = element_rect(fill = "#D8E0E3", color = NA),
-    panel.grid.minor = element_line(size = rel(0.5)),
-    legend.key = element_rect(fill = "#D8E0E3", color = NA),
+    panel.background = ggplot2::element_rect(fill = "#D8E0E3", color = NA),
+    panel.grid.minor = ggplot2::element_line(linewidth = ggplot2::rel(0.5)),
+    legend.key = ggplot2::element_rect(fill = "#D8E0E3", color = NA),
     ...
   )
 }
 
 theme_demigirl <- function(...) {
   ggplot2::theme(
-    panel.background = element_rect(fill = "#E3DADD", color = NA),
-    panel.grid.minor = element_line(size = rel(0.5)),
-    legend.key = element_rect(fill = "#E3DADD", color = NA),
+    panel.background = ggplot2::element_rect(fill = "#E3DADD", color = NA),
+    panel.grid.minor = ggplot2::element_line(linewidth = ggplot2::rel(0.5)),
+    legend.key = ggplot2::element_rect(fill = "#E3DADD", color = NA),
     ...
   )
 }