+++ man/roxygen/templates/name.R
@@ -1,2 +1,0 @@
-#' @param name \code{character(1)}\cr
-#'  Name of the flag the colors are based on.