+++ tests/testthat/test-palettes.R
@@ -15,8 +15,10 @@ test_that("all elements are RGB color codes", {
 test_that("palettes have single-string names", {
   for (name in names(pride_data)) {
     palette <- palette_lgbtq(name)
-    expect_vector(attr(palette, "name", exact = TRUE),
-                  ptype = character(),
-                  size = 1)
+    expect_vector(
+      attr(palette, "name", exact = TRUE),
+      ptype = character(),
+      size = 1
+    )
   }
 })